requirements.txt
django
tzdata
psycopg2-binary
django-leaflet
django-import-export==4.2.1   # Adjust the version based on the latest release
tablib==3.7.0
openpyxl==3.1.2  # For Excel format support
